fix LPPointMake to create LPPoint struct using floats

+++ low_poly_engine/LPEnginePrimitives.h
@@ -33,7 +33,7 @@ typedef struct Vertex {
 
 LPTriangle LPTriangleMake(LPPoint p1, LPPoint p2, LPPoint p3) ;
 
-LPPoint LPPointMake(int16_t x, int16_t y, int16_t z) ;
+LPPoint LPPointMake(float x, float y, float z) ;
 NSString *NSStringFromLPPoint(LPPoint point);
 
 LPPoint pointSubtract(LPPoint p1, LPPoint p2);
